Moved attack roll into attack item where it belongs Start on system settings
