+++ templates/chatcards/attackroll.hbs
@@ -1,5 +1,5 @@
 <span style="font-weight: bold">{{actorName}}</span> {{localize 'MP.attacks'}}
-<span style="font-weight: bold">{{target}}</span> {{localize 'MP.With'}}
+<span style="font-weight: bold">{{targetName}}</span> {{localize 'MP.With'}}
 <span style="font-weight: bold">{{attackName}}</span>
 
 <div class="dice-roll">
