+++ templates/chatcards/savingthrow.hbs
@@ -1,4 +1,5 @@
-{{localize 'MP.Target'}}: <span style="font-weight: bold">{{target}}</span>
+<span style="font-weight: bold">{{stat}}</span> 
+{{localize "MP.SavingThrow"}}:  ({{localize 'MP.Target'}}: <span style="font-weight: bold">{{target}}-)</span>
 
 <div class="dice-roll">
     <div class="dice-result">
